Move numina.core.dal to numina.dal
* And moved NoResultFound exception to numina.exceptions
