+++ numina/dal/__init__.py
@@ -19,7 +19,6 @@
 
 from .absdal import AbsDAL
 from .daliface import DALInterface
-from .daliface import NoResultFound
 from .stored import StoredParameter
 from .stored import StoredProduct
 from .stored import ObservingBlock
